% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nres.R
\name{calculateNres2}
\alias{calculateNres2}
\title{Number of mutated residual genes}
\usage{
calculateNres2(
  patientRes,
  genesRes = c("BCOR", "BCORL1", "CEBPA", "ETNK1", "GATA2", "GNB1", "IDH1", "NF1",
    "PHF6", "PPM1D", "PRPF8", "PTPN11", "SETBP1", "STAG2", "WT1"),
  Nref = 0.388
)
}
\arguments{
\item{patientRes}{a patient \code{data.frame} (with one row and residual genes in columns), or a \code{vector} with residual genes as names attributes. Expected entries are \code{0} (not mutated), \code{1} (mutated), or \code{NA} (missing).}

\item{genesRes}{a \code{vector} containing the names of the residual genes.}

\item{Nref}{the average reference value.}
}
\value{
A \code{vector} with the calculated number under the mean, best, and worst scenarios. If no missing data, all scenarios are equal.
}
\description{
Performs the generalized calculation for the number of mutated residual genes per patient, with a cap at 2 in accordance to observed saturation, and adjusting for missing information.
}
\examples{
vec <- rep(0,15) ; vec[1] <-1 # 1 mutated genes and no missing data
names(vec) <- c("BCOR","BCORL1","CEBPA","ETNK1","GATA2","GNB1","IDH1","NF1","PHF6","PPM1D","PRPF8","PTPN11","SETBP1","STAG2","WT1")
calculateNres2(patientRes=vec)

vec1 <- vec ; vec1[2:3] <- NA # 1 mutated genes and 2 missing
calculateNres2(patientRes=vec1) # scenarios are now different

vec2 <- vec1 ; vec2[4] <- 1 # 2 mutated genes and 2 missing
calculateNres2(patientRes=vec2) # scenarios are equal, because the number is capped at 2 mutated genes
}
