% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPSSMannotate.R
\name{IPSSMannotate}
\alias{IPSSMannotate}
\title{Annotate IPSS-M result}
\usage{
IPSSMannotate(patientResult, range.max = 1)
}
\arguments{
\item{patientResult}{a patient result \code{data.frame}, as the output of \code{IPSSMmain} function.}

\item{range.max}{threshold for the allowed maximal range in IPSS-M risk (worst - best) to confidently report the mean scenario as the main result. Default 1.}
}
\value{
An annotated patient \code{data.frame}, same number of rows/patients as in \code{patientResult}, with additonal columns labelled \code{IPSSMscore}, \code{IPSSMcat}, and \code{Comment}.
}
\description{
Consolidates the IPSS-M results to create single IPSSM score and IPSSM category columns. If the maximal range of IPSS-M risk score (worst-best) is below the allowed limit, then the IPSS-M results under the mean scenario are reported. Of course if no missing data, the IPSS-M values are fully certains.
}
\examples{
dd <- read.csv(system.file("extdata", "IPSSMexample.csv", package = "ipssm"), header = TRUE)
dd.process <- IPSSMprocess(patientInput = dd)
dd.res <- IPSSMmain(patientProcess = dd.process)
dd.annot <- IPSSMannotate(patientResult = dd.res)
print(dd.annot[, c("ID", "IPSSMscore", "IPSSMcat", "Comment")])

}
