% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPSSMprocess.R
\name{IPSSMprocess}
\alias{IPSSMprocess}
\title{Processing of user-defined clinical and molecular variables}
\usage{
IPSSMprocess(
  patientInput,
  genesRes = c("BCOR", "BCORL1", "CEBPA", "ETNK1", "GATA2", "GNB1", "IDH1", "NF1",
    "PHF6", "PPM1D", "PRPF8", "PTPN11", "SETBP1", "STAG2", "WT1"),
  maxvafloh = 0.55,
  Nref = 0.388
)
}
\arguments{
\item{patientInput}{a patient \code{data.frame}, one patient per row, and variables as columns.}

\item{genesRes}{a \code{vector} containing the names of the residual genes.}

\item{Nref}{the average reference value for min(Nres,2) where Nres is the number of mutated residual genes.}
}
\value{
A processed patient \code{data.frame}, same number of rows/patients as in \code{patientInput}, and with the processed variables as additional columns.
}
\description{
Process clinical and molecular user-defined variables to model-based variables.
}
\examples{
dd <- read.csv(system.file("extdata", "IPSSMexample.csv", package = "ipssm"),header=T)
dd.process <- IPSSMprocess(patientInput=dd)
print(dd.process)
}
