% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPSSMread.R
\name{IPSSMread}
\alias{IPSSMread}
\title{Read User Input Data and Validate Format}
\usage{
IPSSMread(path.file, sheet = 1)
}
\arguments{
\item{path.file}{path to the input data file. Allowed formats are .csv, .tsv, .xls, and .xlsx.}

\item{sheet}{sheet number for excel format input file. Default is 1.}
}
\value{
A patient input \code{data.frame}.
}
\description{
Function that reads the user input data and performs some validation for the presence, format, and values of the input data.
}
\examples{
path.file <- system.file("extdata", "IPSSMexample.csv", package = "ipssm")
dd <- IPSSMread(path.file)
print(dd)

path.file2 <- system.file("extdata", "IPSSMexample.xlsx", package = "ipssm")
dd2 <- IPSSMread(path.file2, sheet=2) # equivalent 
}
