% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPSSMread_df.R
\name{IPSSMread_df}
\alias{IPSSMread_df}
\title{Read User Input Data accepts input from data frame}
\usage{
IPSSMread_df(df, calculation = "ALL")
}
\arguments{
\item{patientInput}{a patient \code{data.frame}, one patient per row, and variables as columns.}

\item{genesRes}{a \code{vector} containing the names of the residual genes.}

\item{maxvafloh}{maximum variant allele frequency (vaf) threshold to determine theres Loss of Heterozigosity (LOH).}

\item{Nref}{the average reference value for min(Nres,2) where Nres is the number of mutated residual genes.}
}
\value{
A processed patient \code{data.frame}, same number of rows/patients as in \code{patientInput}, and with the processed variables as additional columns.
}
\description{
Read User Input Data accepts input from data frame
}
