% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPSSMwrapper.R
\name{IPSSMwrapper}
\alias{IPSSMwrapper}
\title{IPSS-M Wrapper}
\usage{
IPSSMwrapper(
  path.file,
  sheet = 1,
  genesRes = c("BCOR", "BCORL1", "CEBPA", "ETNK1", "GATA2", "GNB1", "IDH1", "NF1",
    "PHF6", "PPM1D", "PRPF8", "PTPN11", "SETBP1", "STAG2", "WT1"),
  Nref = 0.388,
  betaValues = c(HB1 = -0.171, TRANSF_PLT100 = -0.222, BLAST5 = 0.352, CYTOVEC = 0.287,
    TP53multi = 1.18, FLT3 = 0.798, MLL_PTD = 0.798, SF3B1_5q = 0.504, NPM1 = 0.43, RUNX1
    = 0.423, NRAS = 0.417, ETV6 = 0.391, IDH2 = 0.379, CBL = 0.295, EZH2 = 0.27, U2AF1 =
    0.247, SRSF2 = 0.239, DNMT3A = 0.221, ASXL1 = 0.213, KRAS = 0.202, SF3B1_alpha =
    -0.0794, nRes2 = 0.231),
  meanValues = c(HB1 = 9.87, TRANSF_PLT100 = 1.41, BLAST5 = 0.922, CYTOVEC = 1.39,
    TP53multi = 0.071, FLT3 = 0.0108, MLL_PTD = 0.0247, SF3B1_5q = 0.0166, NPM1 = 0.0112,
    RUNX1 = 0.126, NRAS = 0.0362, ETV6 = 0.0216, IDH2 = 0.0429, CBL = 0.0473, EZH2 =
    0.0588, U2AF1 = 0.0866, SRSF2 = 0.158, DNMT3A = 0.161, ASXL1 = 0.252, KRAS = 0.0271,
    SF3B1_alpha = 0.186, nRes2 = 0.388),
  bestValues = c(HB1 = 20, TRANSF_PLT100 = 2.5, BLAST5 = 0, CYTOVEC = 0, TP53multi = 0,
    FLT3 = 0, MLL_PTD = 0, SF3B1_5q = 0, NPM1 = 0, RUNX1 = 0, NRAS = 0, ETV6 = 0, IDH2 =
    0, CBL = 0, EZH2 = 0, U2AF1 = 0, SRSF2 = 0, DNMT3A = 0, ASXL1 = 0, KRAS = 0,
    SF3B1_alpha = 1),
  worstValues = c(HB1 = 4, TRANSF_PLT100 = 0, BLAST5 = 4, CYTOVEC = 4, TP53multi = 1,
    FLT3 = 1, MLL_PTD = 1, SF3B1_5q = 1, NPM1 = 1, RUNX1 = 1, NRAS = 1, ETV6 = 1, IDH2 =
    1, CBL = 1, EZH2 = 1, U2AF1 = 1, SRSF2 = 1, DNMT3A = 1, ASXL1 = 1, KRAS = 1,
    SF3B1_alpha = 0),
  rounding = TRUE,
  rounding.digits = 2,
  risk.cutpoints = c(-1.5, -0.5, 0, 0.5, 1.5),
  risk.cat = c("Very Low", "Low", "Moderate Low", "Moderate High", "High", "Very High"),
  range.max = 1
)
}
\arguments{
\item{path.file}{path to the input data file. Allowed formats are .csv, .tsv, .xls, and .xlsx.}

\item{sheet}{sheet number for excel format input file. Default is 1.}

\item{genesRes}{a \code{vector} containing the names of the residual genes. Default is from the IPSS-M publised model.}

\item{Nref}{the average reference value for min(Nres,2) where Nres is the number of mutated residual genes.}

\item{betaValues}{a covariate vector of the model weights. Should have name attributes.}

\item{meanValues}{vector of average values for each covariate. Should have the same names attributes as in \code{betaValues}.}

\item{bestValues}{vector of best values (leading to minimal risk) for each covariate (nRes2 not needed as already taken care of in \code{IPSSMprocess}).}

\item{worstValues}{vector of worst values (leading to maximal risk) for each covariate (nRes2 not needed as already taken care of in \code{IPSSMprocess}).}

\item{rounding}{should the raw IPSS-M risk score be rounded. Default is TRUE.}

\item{rounding.digits}{number of digits for the rounding. Default is 2.}

\item{risk.cutpoints}{cutpoints to be applied to the IPSS-M risk score to create risk categories.}

\item{risk.cat}{names of the IPSS-M risk categories.}

\item{range.max}{threshold for the allowed maximal range in IPSS-M risk (worst - best) to confidently report the mean scenario as the main result. Default 1.}
}
\value{
A patient annoated results \code{data.frame}.
}
\description{
Function that reads, processes, calculates and annotates to output IPSS-M results
}
\examples{
path.file <- system.file("extdata", "IPSSMexample.csv", package = "ipssm")
ddres <- IPSSMwrapper(path.file)
print(ddres)

}
