% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPSSRannotate.R
\name{IPSSRannotate}
\alias{IPSSRannotate}
\title{Annotate IPSS-R result}
\usage{
IPSSRannotate(patientResult, calculation = "ALL")
}
\arguments{
\item{patientResult}{a patient result \code{data.frame}, as the output of \code{IPSSMmain} function.}

\item{range.max}{threshold for the allowed maximal range in IPSS-R risk (worst - best) to confidently report the mean scenario as the main result. Default 1.}
}
\value{
An annotated patient \code{data.frame}, same number of rows/patients as in \code{patientResult}, with additonal columns labelled \code{IPSSRscore}, \code{IPSSRcat}, and \code{Comment}.
}
\description{
Consolidates the IPSS-R results to create single IPSSM score and IPSSR category columns. If the maximal range of IPSS-R risk score (worst-best) is below the allowed limit, then the IPSS-R results under the mean scenario are reported. Of course if no missing data, the IPSS-R values are fully certain.
}
