% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPSSstats.R
\name{IPSSstats}
\alias{IPSSstats}
\title{IPSS Stats}
\usage{
IPSSstats(patientResult)
}
\value{
A patient statistics \code{data.frame}.
}
\description{
Function to add IPSSM and IPSS R stats based on risk_category
}
