% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPSSsurvival.R
\name{IPSSsurvival}
\alias{IPSSsurvival}
\title{IPSS Survival}
\usage{
IPSSsurvival(patientResult)
}
\value{
A patient survival rates \code{data.frame}.
}
\description{
Function to add IPSSM and IPSS R survival stats based on calculated risk categories
}
